Various minor boyscout changes
