+++ src/Components/Pagination/Pagination.tsx
@@ -8,6 +8,7 @@ import styled, { css } from 'styled-components'
 import { Button } from '../Inputs/Button'
 import { themeCommon } from '../../theme/themeCommon'
 import { usePagination, DOTS_R, DOTS_L } from './usePagination'
+import { devices } from '../../theme/themes'
 
 const Container = styled.div`
   width: 100%;
@@ -43,6 +44,9 @@ const Pages = styled.div`
   align-items: center;
   justify-content: center;
   gap: 10px;
+  @media ${devices.tablet} {
+    gap: 5px;
+  }
 `
 interface Props {
   onClick?: () => void
@@ -53,11 +57,11 @@ export interface PaginationProps {
   totalPages: number
   handlePageChange: (page: number) => void
 }
-// TODO: Rinse logic for displaying active page
 
 export const Pagination = ({ currentPage, totalPages, handlePageChange }: PaginationProps) => {
   const paginationRange = usePagination({ currentPage, totalPages })
-
+  const onFirstPage = currentPage - 1 >= 1
+  const onLastPage = currentPage + 1 >= totalPages
   return (
     <Container>
       <Button
@@ -65,8 +69,7 @@ export const Pagination = ({ currentPage, totalPages, handlePageChange }: Pagina
         label="Prev"
         rounded
         onClick={() => {
-          // Rethink else call
-          currentPage - 1 >= 1 && handlePageChange(currentPage - 1)
+          onFirstPage && handlePageChange(currentPage - 1)
         }}
       />
       <Pages>
@@ -92,8 +95,7 @@ export const Pagination = ({ currentPage, totalPages, handlePageChange }: Pagina
         label="Next"
         rounded
         onClick={() => {
-          // Rethink else call
-          currentPage + 1 <= totalPages && handlePageChange(currentPage + 1)
+          onLastPage && handlePageChange(currentPage + 1)
         }}
       />
     </Container>
