+++ src/Pages/Authenticated/Components/Notifications/types.d.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { CardVariant } from './Card'
+
 import { ReactNode } from 'react'
 
 export type CardVariant = 'positive' | 'neutral' | 'negative' | 'update'
